Refactor: Additional Portfolio
