+++ src/controllers/portfolio/portfolio.ts
@@ -1,7 +1,6 @@
 /* eslint-disable import/no-extraneous-dependencies */
 import fetch from 'node-fetch'
 
-import { PINNED_TOKENS } from '../../consts/pinnedTokens'
 import { Account, AccountId } from '../../interfaces/account'
 import { NetworkDescriptor, NetworkId } from '../../interfaces/networkDescriptor'
 /* eslint-disable @typescript-eslint/no-shadow */
@@ -19,6 +18,8 @@ import { CustomToken } from '../../libs/portfolio/customToken'
 import getAccountNetworksWithAssets from '../../libs/portfolio/getNetworksWithAssets'
 import {
   getFlags,
+  getPinnedGasTankTokens,
+  getTotal,
   shouldGetAdditionalPortfolio,
   validateERC20Token
 } from '../../libs/portfolio/helpers'
@@ -205,8 +206,8 @@ export class PortfolioController extends EventEmitter {
       }, {} as AccountState)
 
       if (shouldGetAdditionalPortfolio(selectedAccount)) {
-        state[accountId]['gasTank'] = { isReady: false, isLoading: false, errors: [] }
-        state[accountId]['rewards'] = { isReady: false, isLoading: false, errors: [] }
+        state[accountId].gasTank = { isReady: false, isLoading: false, errors: [] }
+        state[accountId].rewards = { isReady: false, isLoading: false, errors: [] }
       }
 
       this.emitUpdate()
@@ -373,16 +374,6 @@ export class PortfolioController extends EventEmitter {
 
     if (!res) throw new Error('portfolio controller: no res, should never happen')
 
-    const getTotal = (t: any[]) =>
-      t.reduce((cur: any, token: any) => {
-        for (const x of token.priceIn) {
-          cur[x.baseCurrency] =
-            (cur[x.baseCurrency] || 0) + (Number(token.amount) / 10 ** token.decimals) * x.price
-        }
-
-        return cur
-      }, {})
-
     const rewardsTokens = [
       res.data.rewards.xWalletClaimableBalance || [],
       res.data.rewards.walletClaimableBalance || []
@@ -392,6 +383,7 @@ export class PortfolioController extends EventEmitter {
         ...t,
         flags: getFlags(res.data.rewards, 'rewards', t.networkId, t.address)
       }))
+
     accountState.rewards = {
       isReady: true,
       isLoading: false,
@@ -409,62 +401,16 @@ export class PortfolioController extends EventEmitter {
       flags: getFlags(res.data, 'gasTank', t.networkId, t.address)
     }))
 
-    let pinnedGasTankTokens: TokenResult[] = []
-
-    // Don't set pinnedGasTankTokens if the user has > 1 non-zero tokens
-    if (res.data.gasTank.availableGasTankAssets && !hasNonZeroTokens) {
-      const availableGasTankAssets = res.data.gasTank.availableGasTankAssets
-
-      pinnedGasTankTokens = availableGasTankAssets.reduce((acc: TokenResult[], token: any) => {
-        const isGasTankToken = !!gasTankTokens.find(
-          (gasTankToken: TokenResult) =>
-            gasTankToken.symbol.toLowerCase() === token.symbol.toLowerCase()
-        )
-        const isAlreadyPinned = !!acc.find(
-          (accToken) => accToken.symbol.toLowerCase() === token.symbol.toLowerCase()
-        )
-
-        if (isGasTankToken || isAlreadyPinned) return acc
-
-        const correspondingPinnedToken = PINNED_TOKENS.find(
-          (pinnedToken) =>
-            (!('accountId' in pinnedToken) || pinnedToken.accountId === accountId) &&
-            pinnedToken.address === token.address &&
-            pinnedToken.networkId === token.network
-        )
-
-        if (correspondingPinnedToken && correspondingPinnedToken.onGasTank) {
-          acc.push({
-            address: token.address,
-            symbol: token.symbol.toUpperCase(),
-            amount: 0n,
-            networkId: correspondingPinnedToken.networkId,
-            decimals: token.decimals,
-            priceIn: [
-              {
-                baseCurrency: 'usd',
-                price: token.price
-              }
-            ],
-            flags: {
-              rewardsType: null,
-              canTopUpGasTank: true,
-              isFeeToken: true,
-              onGasTank: true
-            }
-          })
-        }
-        return acc
-      }, [])
-    }
-
     accountState.gasTank = {
       isReady: true,
       isLoading: false,
       errors: [],
       result: {
         updateStarted: start,
-        tokens: [...gasTankTokens, ...pinnedGasTankTokens],
+        tokens: [
+          ...gasTankTokens,
+          ...getPinnedGasTankTokens(res.data, hasNonZeroTokens, accountId, gasTankTokens)
+        ],
         total: getTotal(gasTankTokens)
       }
     }
@@ -508,7 +454,11 @@ export class PortfolioController extends EventEmitter {
 
     const accountState = this.latest[accountId]
     const pendingState = this.pending[accountId]
-
+    console.log(
+      'shouldGetAdditionalPortfolio',
+      shouldGetAdditionalPortfolio(selectedAccount)
+      // hasNonZeroTokens
+    )
     if (shouldGetAdditionalPortfolio(selectedAccount)) {
       this.#getAdditionalPortfolio(accountId)
     }
