+++ src/libs/portfolio/helpers.ts
@@ -2,7 +2,8 @@ import { Contract, ZeroAddress } from 'ethers'
 
 import IERC20 from '../../../contracts/compiled/IERC20.json'
 import gasTankFeeTokens from '../../consts/gasTankFeeTokens'
-import { Account } from '../../interfaces/account'
+import { PINNED_TOKENS } from '../../consts/pinnedTokens'
+import { Account, AccountId } from '../../interfaces/account'
 import { NetworkId } from '../../interfaces/networkDescriptor'
 import { RPCProvider } from '../../interfaces/settings'
 import { isSmartAccount } from '../account/account'
@@ -103,4 +104,73 @@ export const shouldGetAdditionalPortfolio = (account: Account) => {
 // otherwise, the token.amount
 export const getTokenAmount = (token: TokenResult): bigint => {
   return typeof token.amountPostSimulation === 'bigint' ? token.amountPostSimulation : token.amount
+}
+
+export const getTotal = (t: any[]) =>
+  t.reduce((cur: any, token: any) => {
+    const localCur = cur
+    if (token.isHidden) return localCur
+    // eslint-disable-next-line no-restricted-syntax
+    for (const x of token.priceIn) {
+      cur[x.baseCurrency] =
+        (cur[x.baseCurrency] || 0) +
+        (Number(getTokenAmount(token)) / 10 ** token.decimals) * x.price
+    }
+
+    return localCur
+  }, {})
+
+export const getPinnedGasTankTokens = (
+  data: any,
+  hasNonZeroTokens: boolean,
+  accountId: AccountId,
+  gasTankTokens: TokenResult[]
+) => {
+  const availableGasTankAssets = data.gasTank?.availableGasTankAssets
+
+  if (!availableGasTankAssets) return []
+  // Don't set pinnedGasTankTokens if the user has > 1 non-zero tokens
+  if (hasNonZeroTokens) return []
+
+  return availableGasTankAssets.reduce((acc: TokenResult[], token: any) => {
+    const isGasTankToken = !!gasTankTokens.find(
+      (gasTankToken: TokenResult) =>
+        gasTankToken.symbol.toLowerCase() === token.symbol.toLowerCase()
+    )
+    const isAlreadyPinned = !!acc.find(
+      (accToken) => accToken.symbol.toLowerCase() === token.symbol.toLowerCase()
+    )
+
+    if (isGasTankToken || isAlreadyPinned) return acc
+
+    const correspondingPinnedToken = PINNED_TOKENS.find(
+      (pinnedToken) =>
+        (!('accountId' in pinnedToken) || pinnedToken.accountId === accountId) &&
+        pinnedToken.address === token.address &&
+        pinnedToken.networkId === token.network
+    )
+
+    if (correspondingPinnedToken && correspondingPinnedToken.onGasTank) {
+      acc.push({
+        address: token.address,
+        symbol: token.symbol.toUpperCase(),
+        amount: 0n,
+        networkId: correspondingPinnedToken.networkId,
+        decimals: token.decimals,
+        priceIn: [
+          {
+            baseCurrency: 'usd',
+            price: token.price
+          }
+        ],
+        flags: {
+          rewardsType: null,
+          canTopUpGasTank: true,
+          isFeeToken: true,
+          onGasTank: true
+        }
+      })
+    }
+    return acc
+  }, [])
 }
